local-travis: add a travis "emulation"
This patch introduces the local-travis, an small util program locally
emulate the remote travis execution. The program will read the travis
yaml file and execute the tasks.

The program is not intended to be manually executed but wrapped by
"make check-coverage". The program uses the env var TRAVIS_CONF to
know where to find the travis yaml file (the makefile will have it all
set for you).

We also use the http_proxy, https_proxy and no_proxy if they are set.
If we're executing a docker command that supports -e flag then these
variables are also handled.

The check-coverage.sh script will also auto update the coverage database
and commit the changes if it detects a coverage increase. In order to
do that check-coverage will only update the database if there's no
unstaged changes in the tree and if it's being executed with
make check-coverage.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>