Setting Log Level should not be fatal
Changing the log level should not abort the installer.
Instead, force to highest or lowest value and log a warning.
This is particularly useful if set via the kernel command line
during debug of an installation we do not want to fail.

Moved setting the output log file name before setting level.

Signed-off-by: Mark D Horn <mark.d.horn@intel.com>