check-coverage: improve error handling and notification
One of the steps check-coverage does is running "make check" and
writing it to a temporary file. This patch improves this step by
detecting the failure and better notifying the user it has failed,
also we show the result to the user if that's the case. Also remove
the left over temporary file case it fails.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>