Improve binary versioning
Now the version string in the binary is set by the
existing tag, plus the sha if changes exists beyond
the tag, plus "dirty" if built with uncommitted changes.

Tagging of releases is now a manual process, but the Makefile
will still print the instructions.

Also saves us creating a PR just to change the version string.

Signed-off-by: Mark D Horn <mark.d.horn@intel.com>