debug: Save config before stating install
Save a copy of the current configuration prior to starting
the installation. This provides a YAML containing the configuration
in case of a crash or failure during the install.
It is placed in the same directory as the clr-installer.log file as
both of these files should be submitted in case of failures.

Signed-off-by: Mark D Horn <mark.d.horn@intel.com>