Disable the root account if Admin set
If any accounts are added and set to have Admin privileges
via sudo, then disable the root account by locking it.

Also lock the root account if SSH Keys are set, but no password.

Added checking if the account exists (system accounts like root)
to use usermod not useradd. Also use getent to find the user's
home directory and not assume /home (not true for root).

Signed-off-by: Mark D Horn <mark.d.horn@intel.com>