Chroot when enabling clr-installer service
When using the --root option with systemctl, libsystemd for the systemd
in the chroot will be resolved on the host.  In cases where the systemd
in the chroot is a different version from the systemd on the host, this
will result in a library not found error.  Need to chroot when eanbling
clr-installer service to avoid this problem.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>