gui: Port of bug fix for Issue: #242
This is the GUI version of the same fix that was implemented
for the TUI of Issue #242.

Only show user selected bundles on the summary menu item for
bundle selection; not core, required bundles.

Make the latest user bundle changes more evident by prepending
them to the list of bundles to install.

Signed-off-by: Mark D Horn <mark.d.horn@intel.com>