Add boot message, disable autolaunch, change PS1
Add a message printed on boot that gives basic information and
instructions to login as root and run clr-installer to install onto this
sytem.

Also change $PS1 on the installer image after logged in to
indicate we're booted into a live image.

Remove clr-installer.service from the Makefile and remove all the
toggling on or off from the yaml files.

Signed-off-by: Brian J Lovin <brian.j.lovin@intel.com>