isoutils: Include custom kernel command line parms
isoutils had a bug which would remove additional kernel command line
paramaters specified during image build time.

To fix this, instead of reading the cmdline file from the rootfs,
read the modified cmdline file from the EFI partition (which
clr-boot-manager fills in for us).

Modify the options as required, and then write in the appropriate
places.

Signed-off-by: Brian J Lovin <brian.j.lovin@intel.com>