Prevent duplicate instances
Create a simple lock file for preventing more than one instances of
the installer from running.

If we have a recoverable crash, like SEGV (accessing a nil pointer)
we will not likely remove the lockfile. For the installer ISO images
the user may have to manually remove the lockfile or reboot.

Signed-off-by: Mark D Horn <mark.d.horn@intel.com>