gui: user: Force the only user to be admin
We want the root account locked on a Desktop installation to
prevent anyone with console access to switching to a virtual
terminal and setting a root password.
If there is an admin user, we will lock the root account.

TODO: When the user manager is added, only one user will be
required to be an admin.

Signed-off-by: Mark D Horn <mark.d.horn@intel.com>