gui: Add kernel options page
New page to set a custom kernel and the associated command line options.

There's a tooltip over the entries that has the same verbiage as is in
the TUI page. There's also a concept of 'default' options, based on what
is currently loaded in the model. This works even on VirtualBox to
mark the LTS kernel as a default.

Fixes #384, Fixes #386

Signed-off-by: Brian J Lovin <brian.j.lovin@intel.com>