tui: Rework of advanced partitioning
Remove the code which was attempting to do the work of many
existing, more reliable disk partitioning tools. The code was
getting overly complex to handle the very high number of options
and corner cases.

The installer will rely on external tools which do a better job.

This will parse the media for advanced partition labels.

Show the user the details of what will change and be used
for the installation.

Signed-off-by: Mark D Horn <mark.d.horn@intel.com>