InstallerYAMLSyntax: fix cmdline bug from example
The kernel-arguments example is missing a pair of quotes for the remove
field-- without the quotes, two removal entries are added to
cmdline-removal.d on installation: "console=ttyS0" and "115200n8".

The quotes ensure that a single cmdline-removal.d entry,
"console=ttyS0,115200n8", is generated at installation time on the
target.

Signed-off-by: Joe Konno <joe.konno@intel.com>