storage: Support for partition labels
Add partition labels to the lsblk parsed data to enable the
advanced media configuration option. An external tool will be
used to set the partition labels to identify media for use.

Check if a file system is present on each label partition, and if
non is found, use the default file system and flag for formatting.

Signed-off-by: Mark D Horn <mark.d.horn@intel.com>