gui: Add network check to install
When a network connection is required to complete an install, but the
installer's network state is not passing, a network check will be
performed. If the network check fails, additional bundles were selected,
and offline installs are supported, the user will be prompted to remove
the additional bundles so that they can continue the install.

Signed-off-by: John Akre <john.w.akre@intel.com>