Add offline install support (#504)
* Support loading configs from chroot

To support offline installs, the bundles listed in the default and
kernel configs from within the installer image must be known. This
change adds functions to read these config files from within a chroot
which can be used to read the config files within the installer image.

Signed-off-by: John Akre <john.w.akre@intel.com>

* Add targetBundles support to config

The targetBundles field in the config file specifies the minimum set of
bundles that the installer image's installer will install to the target
media. To achieve this functionality, a config file is generated and
inserted into the installer image at /var/lib/clr-installer.yaml. This
config file contains the list of bundles specified by targetBundles and
a minimum set of entries based on the config file used to generate the
installer image (keyboard, language, timezone, kernel).

Signed-off-by: John Akre <john.w.akre@intel.com>

* Add offline install support

When creating installer images with offline support, installation
content for a minimum set of bundles will be downloaded to the
/var/lib/clr-installer/offline-content directory inside the installer
image. The offline-content directory will be used instead of the network
to install bundles when it exists.

To generate images with offline support, the --offline command line
argument or the offline config file field can be set.

Signed-off-by: John Akre <john.w.akre@intel.com>

* Don't update to latest version for offline installs

When installing with offline content, the installed version must match
the version of the installer image so that the offline content can be
consumed.

Signed-off-by: John Akre <john.w.akre@intel.com>

* tui: Add network check to install

A network check will be performed when a network connection is required
to install, but hte installer's network state is not passing. If the
network check fails, additional bundles were selected, and offline
installs are supported, the user will be prompted to remove the
additional bundles so that they can continue the install.

Signed-off-by: John Akre <john.w.akre@intel.com>

* gui: Add network check to menu

A network check page was added to the menu. Similarly to the TUI, the
network check page will launch a pop-up window and perform a network
check.

Signed-off-by: John Akre <john.w.akre@intel.com>

* Use indeterminate progress bar for network check

The indeterminate progress bar will continually move until the network
check completes. This type of progress bar makes more sense than the
multistep progress bar that incremented after each network retry.

Signed-off-by: John Akre <john.w.akre@intel.com>

* gui: Add network check to bundle page

The bundle page will peform a network check when the network state is
not passing. Users will be unable to select additional bundles when the
network check fails.

Signed-off-by: John Akre <john.w.akre@intel.com>

* gui: Add network check to install

When a network connection is required to complete an install, but the
installer's network state is not passing, a network check will be
performed. If the network check fails, additional bundles were selected,
and offline installs are supported, the user will be prompted to remove
the additional bundles so that they can continue the install.

Signed-off-by: John Akre <john.w.akre@intel.com>

* tui: Add network check to bundle page

When the network state is not passing, the bundle page will perform a
network check. Users will be unable to select additional bundles when
the network check fails.

Signed-off-by: John Akre <john.w.akre@intel.com>

* gui: Remove precheck page

Network checks are no longer a hard requirement with offline install
support, so the precheck page should not prevent users from advancing.
This change creates a fatal pop-up window for system check failures and
removes the precheck page. Network checks will be performed as needed.

Signed-off-by: John Akre <john.w.akre@intel.com>

* test: reduced args coverage

The Offline and OfflineSet fields were added to args, but test coverage
only exists for Offline. Typically fields like OfflineSet are untested,
so reducing the coverage.

Signed-off-by: John Akre <john.w.akre@intel.com>

* Add installation step to copy offline content

This commit adds an installation step to copy offline update contents to
the install target media. When installing from an ISO image, unsquashfs
is used to decompress offline content directly to the target which is
faster than a file copy.

Signed-off-by: John Akre <john.w.akre@intel.com>

* install: Update indeterminate progress bar

The indeterminate progress bar does not behave as expected when pulsing
twice consecutively. In this case, the progress bar moves very, very
slowly. When the progress bar pulses once, it smoothly moves back and
forth across the screen.

Signed-off-by: John Akre <john.w.akre@intel.com>

* swupd: Specify VerifyWithBundles steps

The VerifyWithBundles function is used to install contents into the
target, download offline contents, and to create the initrd for the ISO.
Now, a prefix string will be added to the output of these commands to
specify which step VerifyWithBundles is running.

Signed-off-by: John Akre <john.w.akre@intel.com>

* docs: Document offline and targetBundles

Signed-off-by: John Akre <john.w.akre@intel.com>