gui: Fix disk config menu description
The saveMedias variable is responsible for selecting between loading the
page or resetting page state on cancel button clicks. When clicking the
confirm button, saveMedias must be set to nil to prevent flipping the
logic for loading the page and cancel button clicks. This logic
inversion led to incorrect menu descriptions for the disk config page.

Signed-off-by: John Akre <john.w.akre@intel.com>