Makefile: Add race condition support to coverage target
The `go build -race` command creates an executable with support for race
condition detection, but does not provide any static analysis. This
change adds race condition detection support to the coverage Makefile
target.

The GUI packages are skipped for the race condition check because
building these packages generates a large number of deprecation warnings
that cannot easily be disabled.

Fixes: #539

Signed-off-by: John Akre <john.w.akre@intel.com>