Refactor main to handle errors without panics
Previously, the main function used panics to exit with a non-zero exit
code and to execute deferred functions (os.Exit() does not execute
deferred functions). To remove the need for panics the main function was
refactored so that it is only responsible for calling the execute function,
logging failures, and returning the exit status.

Fixes: #547

Signed-off-by: John Akre <john.w.akre@intel.com>