Remove kbd and glibc-locale dependencies
The validation errors for the keyboard and language settings have been
changed to warnings. This enables the kbd and glibc-locale bundles to
be set as optional in the clr-installer bundle definition.

When the kbd and glibc-locale bundles are not installed on the host
system, the tui/gui will not populate the list of alternate
keyboards/locales and default settings will be used.

Signed-off-by: John Akre <john.w.akre@intel.com>