pxe: Block service if no config file
Do not start the clr-installer provisioning service if there
was no configuration file supplied on the kernel command line.

As this service has already masked the getty for this tty, no
login prompt will be available on this terminal. The use will
have to switch to a different virtual terminal for login access.

Signed-off-by: Mark D Horn <mark.d.horn@intel.com>